requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet>=3.0.0
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
google-genai>=1.0.0
aiofiles==24.1.0
pypdf2==3.0.1
email-validator>=2.0.0
